Relocate BaseRepositoryFixture.cs, ReadOnlyRepository.cs to Base directory
